implemet JWT

+++ resources/item.py
@@ -2,8 +2,10 @@ from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 
 from db import db
-from models import ItemModel
+from models import ItemModel, StoreModel
 from sqlalchemy.exc import SQLAlchemyError
+
+from flask_jwt_extended import jwt_required, get_jwt
 
 from schemas import ItemSchema, ItemUpdateSchema
 
@@ -11,17 +13,23 @@ blp = Blueprint("Items", __name__, description="Operations on items")
 
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilage required !")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
         return {"message": "Item Deleted !!"}
     
+    @jwt_required()    
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self,item_data,item_id):
@@ -40,19 +48,24 @@ class Item(MethodView):
 @blp.route("/item")
 class ItemList(MethodView):
     
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
     
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
-        item = ItemModel(**item_data)
+        if StoreModel.query.filter(StoreModel.id == item_data["store_id"]).first():
+            
+            item = ItemModel(**item_data)
 
-        try:
-            db.session.add(item)
-            db.session.commit()
-        except SQLAlchemyError:
-            abort(500, message="An error occured while inserting the item")
+            try:
+                db.session.add(item)
+                db.session.commit()
+            except SQLAlchemyError:
+                abort(500, message="An error occured while inserting the item")
 
-        return item
+            return item
+        abort(400, "Store doesn't exist !!")
