+++ resources/store.py
@@ -5,6 +5,8 @@ from db import db
 from models import StoreModel
 from sqlalchemy.exc import SQLAlchemyError , IntegrityError
 
+from flask_jwt_extended import jwt_required, get_jwt
+
 from schemas import StoreSchema
 
 blp = Blueprint("stores", __name__, description="Operations on stores") 
@@ -13,12 +15,18 @@ blp = Blueprint("stores", __name__, description="Operations on stores")
 # using flask MethodView we can craete a class whos methods routes to specific endpoints
 @blp.route("/store/<int:store_id>") # This decorator made the connection between flask-smorest and flask method view
 class Store(MethodView):
+
+    @jwt_required()
     @blp.response(200, StoreSchema)
     def get(self, store_id):
         store= StoreModel.query.get_or_404(store_id)
         return store
 
+    @jwt_required()
     def delete(self, store_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilage required !")
         store= StoreModel.query.get_or_404(store_id) 
         db.session.delete(store)
         db.session.commit()
@@ -33,6 +41,7 @@ class StoreList(MethodView):
     def get(self):
         return StoreModel.query.all()
 
+    @jwt_required()
     @blp.arguments(StoreSchema)
     @blp.response(201, StoreSchema)
     def post(self, store_data):
